moved workspace-info.ts
