+++ src/app/dataview/dataview.component.ts
@@ -3,7 +3,7 @@ import { ActivatedRoute } from '@angular/router';
 import { Component, OnInit, HostListener, ViewChildren, ViewChild, 
          ChangeDetectorRef, AfterViewChecked } from '@angular/core';
 import { DataloaderService } from '../data-loader/data-loader.service';
-import { WorkspaceInfo, TypeMap, LabelScheme } from '../data-loader/workspace-info';
+import { WorkspaceInfo, TypeMap, LabelScheme } from '../data-loader/info/workspace-info';
 import { DataInfo } from "../data-loader/info/data-info";
 import { SettingsService } from '../settings/settings.service';
 import { Label, LabelStream} from './labelstreams/labelstream';
