+++ src/app/home/workspace/workspace.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { WorkspaceInfo, LabelScheme } from '../../data-loader/workspace-info';
+import { WorkspaceInfo, LabelScheme } from '../../data-loader/info/workspace-info';
 import { DataInfo } from "../../data-loader/info/data-info";
 
 // #region [Interfaces]
